fix(merge/queue): ensure context is updated on merged PR
If GitHub is laggy, the next /pull/XXX after a PR get merged may not be
marked as merged. Since the merge queue code relies on this to known if
something unexpectedly merged in meantimes, we must wait the attributes
merged change.

Fixes MRGFY-1272

+++ mergify_engine/context.py
@@ -37,6 +37,7 @@ import jinja2.sandbox
 import jinja2.utils
 import markdownify
 import msgpack
+import tenacity
 
 from mergify_engine import cache
 from mergify_engine import check_api
@@ -1693,13 +1694,20 @@ class Context(object):
             and self.pull["mergeable"] is not None
         )
 
-    async def update(self) -> None:
-        # TODO(sileht): Remove me,
+    @tenacity.retry(
+        wait=tenacity.wait_exponential(multiplier=0.2),
+        stop=tenacity.stop_after_attempt(5),
+        retry=tenacity.retry_never,
+        reraise=True,
+    )
+    async def update(self, wait_merged: bool = False) -> None:
         # Don't use it, because consolidated data are not updated after that.
-        # Only used by merge action for posting an update report after rebase.
+        # Only used by merge/queue action for posting an update report after rebase.
         self.pull = await self.client.item(
             f"{self.base_url}/pulls/{self.pull['number']}"
         )
+        if wait_merged and not self.pull["merged"]:
+            raise tenacity.TryAgain
         self._caches.pull_check_runs.delete()
 
     async def _get_external_parents(self) -> typing.Set[github_types.SHAType]:
